CloudFunctions/get_clean_upload_data/requirements.txt
google-cloud-storage
yfinance
